Finish of code to write out the ICC profile information into the clist.  A slight change from the original design.  With this change, the ICC profile data is written out into the cfile and the offset information of the profile is stored along with the profile hash code in a table.  When the clist writer completes, this table (or a serialized version of it) is written into the cfile for band max_band + ICC_OFFSET.   When the reader is executed, it will check the bfile for the existence of this entry at that band.  Since it is written out at the end of the process, that query will be made backwards through the bfile.
git-svn-id: http://svn.ghostscript.com/ghostscript/branches/icc_work@10239 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/base/gstparam.h
@@ -149,7 +149,7 @@ typedef struct gx_transparency_mask_params_s {
 	     sizeof(((gs_pdf14trans_params_t *)0)->GrayBackground))	     
 #define MAX_CLIST_TRANSPARENCY_COMPOSITOR_SIZE (MAX_CLIST_TRANSPARENCY_BUFFER_SIZE + \
 	     sizeof(((gs_pdf14trans_params_t *)0)->transfer_fn)) + \
-             sizeof(int) /* ICC band information */
+             sizeof(int64_t) /* ICC band information */
 #define MAX_CLIST_COMPOSITOR_SIZE MAX_CLIST_TRANSPARENCY_COMPOSITOR_SIZE
 
 /* Select the opacity or shape parameters. */
